refactor(tinacms): creates media directory
